Remove multiple filters, use structuredClone
