+++ services/ui-src/src/components/tables/SortableNaaarStandardsTable.test.tsx
@@ -2,9 +2,6 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 // components
 import { SortableNaaarStandardsTable } from "components";
-import { mapNaaarStandardsData } from "./SortableNaaarStandardsTable";
-// types
-import { EntityShape } from "types";
 // utils
 import { useStore } from "utils";
 import {
@@ -61,54 +58,5 @@ describe("<SortableNaaarStandardsTable />", () => {
     });
 
     testA11y(sortableTableComponent);
-  });
-
-  describe("mapNaaarStandardsData()", () => {
-    test("returns correct data shape", () => {
-      const tableData = mapNaaarStandardsData(mockNaaarStandards);
-      const expectedData = [
-        {
-          count: 1,
-          provider: "Mock Provider; Mock Other Provider",
-          standardType: "Mock Standard Type",
-          description: "Mock Description",
-          analysisMethods: "Mock Method 1, Mock Method 2",
-          population: "Mock Population",
-          region: "Mock Other Region",
-          entity: mockNaaarStandards[0],
-        },
-      ];
-
-      expect(tableData).toEqual(expectedData);
-    });
-
-    test("omits undefined properties", () => {
-      const incompleteData: EntityShape[] = [
-        {
-          id: "mockStandard",
-          standard_coreProviderTypeCoveredByStandard: [
-            { key: "mockProviderType", value: "Mock Provider" },
-          ],
-          standard_standardType: [
-            { key: "mockStandardType", value: "Mock Standard Type" },
-          ],
-          standard_populationCoveredByStandard: [
-            { key: "mockPopulation", value: "Mock Population" },
-          ],
-        },
-      ];
-      const tableData = mapNaaarStandardsData(incompleteData);
-      const expectedData = [
-        {
-          count: 1,
-          provider: "Mock Provider",
-          standardType: "Mock Standard Type",
-          population: "Mock Population",
-          entity: incompleteData[0],
-        },
-      ];
-
-      expect(tableData).toEqual(expectedData);
-    });
   });
 });
